% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid.R
\name{add_grid}
\alias{add_grid}
\title{Add grid/buffer}
\usage{
add_grid(g = NULL, box = NULL, nx = NULL, n = 2)
}
\arguments{
\item{g}{A ggplot objects on which to add the grid}

\item{box}{A boundary box object to get the box extent}

\item{nx}{A number defining the grid interval in metre for x and y}

\item{n}{A number to set the border of the grid from the edge of the plot}
}
\value{
A ggplot2 object
}
\description{
Add a grid and add buffers on a map to assess the area of same land cover
type.
}
\examples{
# Get station metadata
stn <- get_latlon_frost(stationid = 18700)

# Get coordinates and define boundary box
centre <- sf::st_coordinates(stn)
box <- make_bbox(centre, dx = 1600)

# Plot map tile
g <- plot_tile_station(stn, box, tile_name = "ortofoto")

# Add grid and buffer
nx <- 200
n <- 2
g <- add_grid(g, box, nx, n)
g <- add_buffer(g, centre, 300, 1000, nx, n)
g

}
