% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_custom_colours.R
\docType{data}
\name{colour_equipmenttype}
\alias{colour_equipmenttype}
\title{Vector of custom colors for equipmenttype type classification}
\format{
An object of class \code{character} of length 11.
}
\usage{
colour_equipmenttype
}
\description{
Vector of custom colors for equipmenttype type classification
}
\examples{
colour_equipmenttype

}
\keyword{datasets}
