% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_custom_colours.R
\docType{data}
\name{colour_range}
\alias{colour_range}
\title{Colours for scale_fill_gradientn() in \code{"plot_station_horizon_range"}}
\format{
An object of class \code{list} of length 2.
}
\usage{
colour_range
}
\description{
Colour palette is attributed for a set of range values in metre and
changes at 100 m corresponding to the default distance to download
high resolution DEMs
}
\examples{
colour_range

## Generated from
# c(rev(RColorBrewer::brewer.pal(6, "Paired")),
#       RColorBrewer::brewer.pal(4, "Purples"))

}
\keyword{datasets}
