% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_class.R
\name{compute_class}
\alias{compute_class}
\title{Compute station class}
\usage{
compute_class(df.in, test.type = "WMO", f.plot = TRUE)
}
\arguments{
\item{df.in}{A data.frame with land type distribution from \code{"compute_landtype_distance"}}

\item{test.type}{A string defining a type of tests to apply: "WMO" or "Met"}

\item{f.plot}{A boolean whether to plot figures}
}
\value{
A class
}
\description{
Assess the class of a station from the type of surrounding land cover, slope,
vegetation and shadowing as potential heat sources affecting sensors and
using the requirements set by WMO and Met
}
\examples{
# Get station metadata
stn <- get_latlon_frost(stationid=18700)

# Compute land type
landtype <- compute_landtype(stn, dx=100, f.plot=TRUE)

# Compute land type distance to station
landtype_dist <- compute_landtype_distance(stn, landtype, dx=100, f.plot=TRUE)

# Compute class
compute_class(landtype_dist, test.type="WMO", f.plot=TRUE)


}
\references{
\url{https://community.wmo.int/en/activity-areas/imop/siting-classification}
}
