% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_horizon.R
\name{compute_horizon}
\alias{compute_horizon}
\title{Horizon height at a station}
\usage{
compute_horizon(
  centre = NULL,
  dem = NULL,
  level = 2,
  step = 10,
  f.plot.polygon = F
)
}
\arguments{
\item{centre}{An array of two coordinates in UTM 33 (epsg:25833)}

\item{dem}{A SpatRaster of a digital elevation/surface model in UTM 33 (epsg:25833)}

\item{level}{A height above the ground of the sensor in meters, if level is 0, its value is set to 2 m as default}

\item{step}{An interval in degrees at which the horizon will be computed, default is every 10 deg.}

\item{f.plot.polygon}{A boolean setting boundary values to -20 deg to facilitate plotting as polygon}
}
\value{
horizon_height in degrees and azimuth angle at which the horizon is computed in degrees
}
\description{
Compute horizon height in degrees from a station location in UTM 33
}
\examples{
compute_horizon(stn.centre,dem,level=stn.level,step=.01,f.plot.polygon=T)
compute_horizon(stn.centre,dem,level=stn.level,step=.01,f.plot.polygon=T)

}
\references{
\url{https://grass.osgeo.org/grass83/manuals/r.horizon.html}
}
