% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_horizon_rollmean.R
\name{compute_horizon_rollmean}
\alias{compute_horizon_rollmean}
\title{Apply Hourly Rolling Mean to Horizon Height}
\usage{
compute_horizon_rollmean(stn = NULL, horizon = NULL)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_latlon_frost"}}

\item{horizon}{A data.frame with horizon distribution from
\code{"compute_horizon_max"}}
}
\value{
an array of horizon height in degrees
}
\description{
Compute azimuth interval for a theoretical hour based on sun position and
apply a rolling mean with this computed window size on horizon height.
}
\examples{
# Load the station metadata including location and level
stn <- get_metadata_frost(stationid = 18700, paramid = 211)

# Compute maximum horizon
horizon <- compute_horizon_max(stn, step = 1, f_plot_polygon = FALSE)

# Apply rolling mean to horizon height with a computed hourly window
horizon["horizon_mean"] <- compute_horizon_rollmean(stn, horizon)

}
\references{
\url{https://github.com/adokter/suntools/}
}
