% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_landcover.R
\name{compute_landcover}
\alias{compute_landcover}
\title{Compute landcover types}
\usage{
compute_landcover(stn = NULL, dx = 100)
}
\arguments{
\item{stn}{A SpatVector with station attributes from \code{"get_latlon_frost"}}

\item{dx}{A distance in metre or radius defining the extent of the bounding box from the centre point, default `100` metres}
}
\value{
A SpatVector of land cover types
}
\description{
Compute land cover types around a station based on retrieved topographic data
(buildings, roads, water) and digital terrain/surface models (to compute
vegetation height and classify: grass, bush, tree) from Kartverket
}
\examples{
# Get station metadata
stn <- get_latlon_frost(stationid=18700)

# Compute land cover
compute_landcover(stn, dx=100)

}
\references{
\url{https://kartkatalog.geonorge.no/metadata/fkb-wms/84178e68-f40d-4bb4-b9f6-9bfdee2bcc7a}
}
