% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_landtype.R
\name{compute_landtype}
\alias{compute_landtype}
\title{Compute land cover types}
\usage{
compute_landtype(stn = NULL, dx = 100, resx = 1, f_plot = FALSE)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_latlon_frost"}}

\item{dx}{A distance in metre or radius defining the extent of the
bounding box from the centre point}

\item{resx}{A horizontal resolution in metre}

\item{f_plot}{A boolean whether to plot the land cover types}
}
\value{
A SpatVector of land cover types
}
\description{
Compute land cover types around a station based on retrieved topographic data
(buildings, roads, water) and digital terrain/surface models (to compute
vegetation height and classify: grass, bush, tree) from Kartverket. The
horizontal resolution of the digital elevation models defines the resolution
of the tile of the topographic data.
}
\examples{
# Get station metadata
stn <- get_metadata_frost(stationid=18700)

# Compute land cover
compute_landtype(stn, dx=100, resx=1, f_plot=TRUE)

}
\references{
\url{https://kartkatalog.geonorge.no/metadata/fkb-wms/
84178e68-f40d-4bb4-b9f6-9bfdee2bcc7a}
}
