% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_landtype_horizon_class.R
\name{compute_landtype_horizon_class}
\alias{compute_landtype_horizon_class}
\title{Pipeline to compute siting class}
\usage{
compute_landtype_horizon_class(stn)
}
\arguments{
\item{stn}{A SpatVector with station attribute `stationid` from
\code{"get_latlon_frost"}}
}
\value{
None
}
\description{
The pipeline derives the exposure of a station based WMO/met.no siting
classification and includes four input functions:
\code{"download_dem_kartverket"}, \code{"compute_landtype"},
\code{"compute_landtype_distance"} and \code{"compute_horizon_max"}.
They provide the necessary input for \code{"compute_class"}
}
\examples{
# Pipeline to compute necessary input for deriving temperature siting class
# compute_landtype_horizon_class(stn)

}
