% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_skyviewfactor.R
\name{compute_skyviewfactor}
\alias{compute_skyviewfactor}
\title{Sky View Factor (SVF)}
\usage{
compute_skyviewfactor(horizon = NULL)
}
\arguments{
\item{horizon}{An array of horizon height in degrees from
\code{"compute_horizon"}}
}
\value{
Estimated sky view factor, 1 defines an open sky view and 0 a
       totally obstructed sky view
}
\description{
Compute Sky View Factor from horizon data
}
\examples{
# Synthetic example of horizon, see \code{"compute_horizon"}
horizon <- data.frame(azimuth = seq(0, 330, 30),
                     horizon_height = c(7.9, 17.5, 3.8, 7.8, 10.9, 20.1,
                                        23.8, 13.7, 7.0, 7.4, 21.2, 16.2))
compute_skyviewfactor(horizon)

}
\references{
\url{https://github.com/OSGeo/grass-addons/blob/grass8/src/raster/r.skyview/r.skyview.py}
}
