% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dem_kartverket.R
\name{download_dem_kartverket}
\alias{download_dem_kartverket}
\title{Download digital elevation models from Kartverket's WCS API}
\usage{
download_dem_kartverket(
  stationid = NULL,
  centre = NULL,
  name = "dom",
  dx = 100,
  resx = dx/100,
  path = "data/dem",
  f.overwrite = FALSE
)
}
\arguments{
\item{stationid}{A station ID used for the DEM file name}

\item{centre}{A coordinate array (i.e. `c(x, y)`) of the station in UTM 33 (i.e. epsg:25833)}

\item{name}{A name of the DEM to download, either "dtm" a terrain model or the default "dom" a surface model}

\item{dx}{A distance in metre or radius defining the extent of the bounding box from the centre point, default `100` metres}

\item{resx}{A horizontal resolution in metre, default is `dx/100` if greater than `1` metre}

\item{path}{A directory path defining where will be saved the data, default path is "data/dem"}

\item{f.overwrite}{A boolean whether the DEM file should be overwritten, default `FALSE`}
}
\value{
A Digital Elevation Model
}
\description{
Define a GetCapabilities request URL using OSW4R and Kartverket's Web Coverage Service
that downloads a DEM from a bounding box and a DEM type (i.e. "dom" or "dtm").
The bounding box is centered to a parsed location and a parsed radius set the
extent. The downloaded DEM is a SpatRaster object. If the DEM file already
exists, it is loaded by default unless f.overwrite is set to TRUE
}
\examples{
require(sf)

# Define parameters
stationid <- 18700
stn <- get_latlon_frost(stationid)
centre <- sf::st_coordinates(stn)
path   <- "data/dem"

# Load data using ows4R ## WCSClient$new() getCapabilities()
dem    <- download_dem_kartverket(stationid,centre,name="dtm",dx=100,resx=1,path=path)
dsm    <- download_dem_kartverket(stationid,centre,name="dom",dx=100,resx=1,path=path)
demkm  <- download_dem_kartverket(stationid,centre,name="dtm",dx=20e3,resx=20,path=path)

}
\references{
\url{https://kartkatalog.geonorge.no/metadata/nasjonal-hoeydemodell-digital-terrengmodell-25833-wcs/0f0a0f38-00c4-4213-a9e5-2d861dc4abb0}

\url{https://kartkatalog.geonorge.no/metadata/nasjonal-hoeydemodell-digital-overflatemodell-25833-wcs/e36ea427-13a1-4d7c-be82-977068dfc3e3}

\url{https://cran.r-project.org/web/packages/ows4R/vignettes/wcs.html}
}
