% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dem_kartverket_ogc.R
\name{download_dem_kartverket_ogc}
\alias{download_dem_kartverket_ogc}
\title{Download DEM from Kartverket's NEWEST OGC API}
\usage{
download_dem_kartverket_ogc(
  stationid = 18703,
  centre,
  name = "dom",
  dx = 100,
  resx = 1,
  f.OGC = TRUE
)
}
\arguments{
\item{stationid}{A station ID used for the DEM file name}

\item{centre}{A coordinate array (i.e. `c(x, y)`) of the station in UTM 33 (i.e. `epsg:25833`)}

\item{name}{A name of the DEM to download, either `"dtm"` a terrain model or the default `"dom"` a surface model}

\item{dx}{A distance in metre or radius defining the extent of the bounding box from the centre point, default `100` metres}

\item{resx}{A horizontal resolution in metre, default is `dx/100` if greater than `1` metre}

\item{f.OGC}{A boolean default is `TRUE`}
}
\value{
A DEM
}
\description{
Download DEM from Kartverket's NEWEST OGC API
}
\examples{
# To do
}
