% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{extract_resp}
\alias{extract_resp}
\title{Loop through request responses}
\usage{
extract_resp(resp)
}
\arguments{
\item{resp}{A request response}
}
\value{
A matrix
}
\description{
Loop through request responses from pagination to extract metadata
}
\examples{
#No example

}
