% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_custom_colours.R
\docType{data}
\name{fill_corine}
\alias{fill_corine}
\title{Vector of CORINE land cover colors and types}
\format{
An object of class \code{character} of length 45.
}
\usage{
fill_corine
}
\description{
Vector of CORINE land cover colors and types
}
\examples{
fill_corine
# col2rgb(fill_corine)

}
\references{
\url{https://custom-scripts.sentinel-hub.com/copernicus_services/corine/corine_land_cover/}
}
\keyword{datasets}
