% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_custom_colours.R
\docType{data}
\name{fill_landtype}
\alias{fill_landtype}
\title{Vector of custom colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
fill_landtype
}
\description{
Vector of custom colors
}
\examples{
fill_landtype

}
\keyword{datasets}
