% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_pagination.R
\name{format_resp}
\alias{format_resp}
\title{Extract and format JSON metadata}
\usage{
format_resp(res_1)
}
\arguments{
\item{res_1}{A JSON object from an individual station or parameterid}
}
\value{
A data.frame
}
\description{
Extract and format JSON metadata
}
\examples{
#No example
}
