% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_frost.R
\name{get_metadata_frost}
\alias{get_metadata_frost}
\title{Get station and sensor metadata from Frost API}
\usage{
get_metadata_frost(
  stationid = 18700,
  paramid = NULL,
  dx = 100,
  resx = 1,
  path = sprintf("output/\%i", stationid)
)
}
\arguments{
\item{stationid}{A station number as integer and defined by met.no}

\item{paramid}{A parameter number as integer and defined by met.no}

\item{dx}{A distance in metre or radius defining the extent of the
bounding box from the centre point}

\item{resx}{A horizontal resolution in metre}

\item{path}{A directory path defining where will be saved project plots/data}
}
\value{
Station metadata
}
\description{
Fetch station metadata based on station number & parameter id from Frost v1.
The function gets station name, location and sensor details such as level,
exposure and performance
}
\examples{
get_metadata_frost(stationid = 18700)
get_metadata_frost(stationid = 18700, paramid = 211)
get_metadata_frost(stationid = 18700, paramid = 211, dx = 100, resx = 1)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
