% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_pagination.R
\name{get_metadata_pagination}
\alias{get_metadata_pagination}
\title{Get station metadata/coordinates from Frost API}
\usage{
get_metadata_pagination(
  country = "norge",
  paramid = 211,
  timestamp = "1000-01-01T00:00:00Z/2100-01-01T00:00:00Z",
  hdrshow =
    "{\\"id\\":{},\\"extra\\":{\\"station\\":{\\"location\\":[],\\"alternateids\\":[]}}}",
  path = sprintf("output/stn_\%s_\%i", country, paramid)
)
}
\arguments{
\item{country}{A country name as string and defined by met.no}

\item{paramid}{A parameter number as integer and defined by met.no}

\item{timestamp}{A time range as timestamp or string "latest"}

\item{hdrshow}{A JSON string containing metadata structure to extract}

\item{path}{A path where will be saved the metadata as SpatVector `.gpkg`}
}
\value{
Stations metadata
}
\description{
Fetch stations metadata based on country, parameter id, ... using Frost v1.
To handle such a large request output, a pagination protocol is used.
The function gets station names, locations and alternate ids.
}
\examples{
get_metadata_pagination(country = "norge", paramid = 211)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
