% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wms_tile.R
\name{get_wms_tile}
\alias{get_wms_tile}
\title{Get Web Mapping Service data tiles}
\usage{
get_wms_tile(box = NULL, layer = "CORINE_Land_Cover_2012", px = 500)
}
\arguments{
\item{box}{A SpatExtent defining the area to plot}

\item{layer}{A name of a layer provided by WMS such as "ar5", "CORINE_Land_Cover_2012" (default) and "Urban_Atlas_Lu_Lc_2012"}

\item{px}{A pixel number that defines the resolution of the image/tile, default is 500 px}
}
\value{
A map tile
}
\description{
Fetch map tiles from publicly available WMS released by Nibio
}
\examples{
tile <- get_wms_tile(box, layer = "ar5")
tile <- get_wms_tile(box, layer = "CORINE_Land_Cover_2012" )
tile <- get_wms_tile(box, layer = "Urban_Atlas_Lu_Lc_2012" )

}
\references{
\url{https://nibio.no/tjenester/wms-tjenester}
}
