% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_pagination.R
\name{json_to_df}
\alias{json_to_df}
\title{Convert JSON to dataframe}
\usage{
json_to_df(json)
}
\arguments{
\item{json}{A JSON object}
}
\value{
A data.frame
}
\description{
Convert JSON to dataframe
}
\examples{
#No example
}
