% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_ar5.R
\name{load_data_ar5}
\alias{load_data_ar5}
\title{Load FKB-AR5 from vector files stored locally (restricted access to download)}
\usage{
load_data_ar5(box = NULL, f.wms = TRUE, layer = "Arealtype", px = 500)
}
\arguments{
\item{box}{A SpatExtent defining the area to plot}

\item{f.wms}{A boolean to switch to WMS image tile instead of loading file, default TRUE (deprecated)}

\item{layer}{A name of a layer provided by WMS. For "ar5", it is "Arealtype" as default}

\item{px}{A pixel number that defines the resolution of the image/tile, default is 500 px}
}
\value{
A map tile
}
\description{
Load vector FKB-AR5 files to assess area type aka land cover, alternatively
get WMS tile (deprecated, use \code{"get_wms_tile"})
}
\examples{
ar5 <- load_data_ar5(box,f.wms=F)

}
\references{
\url{https://kartkatalog.geonorge.no/metadata/fkb-ar5/166382b4-82d6-4ea9-a68e-6fd0c87bf788}
}
