% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_pagination.R
\name{next_req}
\alias{next_req}
\title{Pagination protocol for API request}
\usage{
next_req(resp, req)
}
\arguments{
\item{resp}{A request response}

\item{req}{A request to update}
}
\value{
Updated request headers
}
\description{
Fetch URL request headers to reassign them with the next request infos
}
\examples{
#No example

}
