% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dem_rayshader.R
\name{plot_dem_rayshader}
\alias{plot_dem_rayshader}
\title{Plot digital surface/elevation model in 3D with shadow rendering}
\usage{
plot_dem_rayshader(stn = NULL, dsm = NULL, path = NULL)
}
\arguments{
\item{stn}{A SpatVector with station attributes from \code{"get_latlon_frost"}}

\item{dsm}{A SpatRaster of a digital surface model around the station, expected radius is 100 m}

\item{path}{A string path that defines where to save the plot, if NULL (default) the plot is printed on-screen and not saved}
}
\value{
A rendered image
}
\description{
Use rayshader to plot surface elevation model near a weather station from
four cardinal angles
}
\examples{
plot_dem_rayshader(stn,dsm, path=path)

}
\references{
\url{https://www.rayshader.com/}
}
