% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_grid.R
\name{plot_station_grid}
\alias{plot_station_grid}
\title{Plot station with Ortophoto, grid and buffers}
\usage{
plot_station_grid(stn = NULL, tile_name = "ortofoto", path = NULL)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_latlon_frost"}}

\item{tile_name}{A string naming a tile to pass to \code{"get_tile_wms"}}

\item{path}{A directory path defining where will be saved the plots,
if path is NULL the plots are printed to the console}
}
\value{
A ggplot2 object
}
\description{
Overlay a grid and two buffers to count the area of land cover types and
assess heat source proximity to a weather station and their distribution.
Four different scales are plotted: 10 m, 30 m, 100 m and 1000 m.
}
\examples{
# Get station metadata
stn    <- get_metadata_frost(stationid = 18700)

# Plot grid and buffers on four different scales
plot_station_grid(stn, path = "plot/map")

}
