% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_horizon_max.R
\name{plot_station_horizon_max}
\alias{plot_station_horizon_max}
\title{Plot max horizon with sun height}
\usage{
plot_station_horizon_max(stn = NULL, horizon_max = NULL, path = stn$path)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_metadata_frost"}}

\item{horizon_max}{A data.frame from \code{"compute_horizon_max"}}

\item{path}{A directory path defining where will be saved the plots,
if path is NULL the plots are printed to the console}
}
\value{
A figure
}
\description{
Plot sun position and maximum horizon from local terrain model, local surface model
and regional digital terrain model with infos from the weather station.
}
\examples{
# Load the station metadata and location
stn <- get_metadata_frost(stationid = 18700,
                          paramid = 211,
                          dx = 100,
                          resx = 1,
                          path = "plot/horizon")

horizon_max <- compute_horizon_max(stn,
                                   step = 1,
                                   f_plot_polygon = TRUE)

# Plot sun diagram and save
plot_station_horizon_max(stn, horizon_max)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
