% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_horizon_range.R
\name{plot_station_horizon_range}
\alias{plot_station_horizon_range}
\title{Plot max horizon and range from the station}
\usage{
plot_station_horizon_range(stn = NULL, path = stn$path)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_metadata_frost"}}

\item{path}{A directory path defining where will be saved the plots,
if path is NULL the plots are printed to the console}
}
\value{
A figure
}
\description{
Plot maximum horizon and its range based on the local terrain model, the
local surface model and the regional digital terrain model with infos
from the weather station.
}
\examples{
# Load the station metadata and location
stn <- get_metadata_frost(stationid = 18700, path = "plot/horizon")

# Plot sun diagram and save
plot_station_horizon_range(stn)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
