% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_raster_to_vector.R
\name{raster_to_vector}
\alias{raster_to_vector}
\title{Convert SpatRaster to SpatVector}
\usage{
raster_to_vector(raster, id = "undefined", mask_thr = 255, f_plot = FALSE)
}
\arguments{
\item{raster}{A SpatRaster image or mask}

\item{id}{A string or number defining a common factor for all generated
polygons}

\item{mask_thr}{A number or boolean defining a background value to remove}

\item{f_plot}{A boolean whether to plot the vectorisation result}
}
\value{
A SpatVector
}
\description{
Convert a raster to polygon, remove raster background from vector
(white for image, FALSE for mask),
aggregate single polygons and set vector values as a single id.
}
\examples{
# Load station
stn <- get_metadata_frost(stationid = 18700)

# Construct box with 200 m radius to extract WMS tile
box <- make_bbox(stn, dx = 200)

# Load a tile
building  <- get_tile_wms(box, layer = "bygning", px = 400)

# Convert raster tile to vector landcover
v_building <- raster_to_vector(building, id = "building", mask_thr = 255)

}
\references{
\url{https://rspatial.r-universe.dev/terra/
doc/manual.html#as.polygons}
}
