% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{stationid_array}
\alias{stationid_array}
\title{Produce sun diagram and maps at the station}
\format{
An object of class \code{numeric} of length 14.
}
\usage{
stationid_array
}
\value{
plots or images of the sun diagram and maps
}
\description{
Run plot_station_siting_context
}
\references{
\url{https://community.wmo.int/en/activity-areas/imop/siting-classification#MembersTools}
}
\keyword{datasets}
