% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_title.R
\name{str_to_title}
\alias{str_to_title}
\title{String to Title format}
\usage{
str_to_title(str)
}
\arguments{
\item{str}{A string}
}
\value{
A capitalised string
}
\description{
Capitalise first letters of each word to match title format
}
\examples{
str <- "OSLO - BLINDERN"
str_to_title(str)

}
\references{
\url{https://stackoverflow.com/questions/6364783/
capitalize-the-first-letter-of-both-words-in-a-two-word-string}
}
