% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_class.R
\name{compute_class}
\alias{compute_class}
\title{Compute station class}
\usage{
compute_class(
  land = landtype_dist,
  horizon = horizon_max,
  dem = dem,
  test_type = "WMO",
  f_plot = TRUE,
  path = path
)
}
\arguments{
\item{land}{A data.frame with land type distribution from
\code{"compute_landtype_distance"}}

\item{horizon}{A data.frame with horizon distribution from
\code{"compute_horizon"}}

\item{dem}{A SpatRaster terrain model}

\item{test_type}{A string defining a type of tests to apply: "WMO" or "Met"}

\item{f_plot}{A boolean whether to plot figures}
}
\value{
A class
}
\description{
Assess the class of a station from the type of surrounding land cover, slope,
vegetation and shadowing as potential heat sources affecting sensors and
using the requirements set by WMO and Met
}
\examples{
# Get station metadata
stn <- get_metadata_frost(stationid=18700)

# Parameters
dx   <- 100
resx <- 1

# Load a digital elevation model
dem   <- download_dem_kartverket(stn, name = "dtm", dx, resx)

# Compute land type
landtype <- compute_landtype(stn, dx, resx, f_plot = TRUE)

# Compute land type distance to station
landtype_dist <- compute_landtype_distance(stn, landtype, dx, resx,
                                           f_plot = TRUE)

# Compute maximum horizon
horizon_max <- compute_horizon_max(stn, dx, resx, step = 1,
                                   f_plot_polygon = FALSE)

# Compute class
compute_class(landtype_dist, horizon_max, dem, test_type = "WMO",
              f_plot = TRUE)


}
\references{
\url{https://community.wmo.int/en/activity-areas/
imop/siting-classification}
}
