% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_horizon.R
\name{compute_horizon}
\alias{compute_horizon}
\title{Horizon height at a station}
\usage{
compute_horizon(
  centre = NULL,
  dem = NULL,
  level = 2,
  step = 10,
  f.plot.polygon = FALSE
)
}
\arguments{
\item{centre}{An array of two coordinates in UTM 33 (epsg:25833)}

\item{dem}{A SpatRaster of a digital elevation/surface model in UTM 33 (epsg:25833)}

\item{level}{A height above the ground of the sensor in metres, level 0 is set to 2 metres}

\item{step}{An interval in degrees at which the horizon will be computed, default is every 10 deg.}

\item{f.plot.polygon}{A boolean setting boundary values to -20 deg to facilitate plotting as polygon}
}
\value{
A dataframe with `horizon_height` in degrees and `azimuth` angle in degrees at which the horizon is computed in degrees
}
\description{
Compute horizon height in degrees from a station location in UTM 33
}
\examples{
require(sf)

# Load the station metadata including location and level
stn <- get_latlon_frost(stationid=18700, paramid=211)
stn.id      <- stn$id.stationid
stn.centre  <- sf::st_coordinates(stn)
stn.level   <- stn$id.level

# Load a digital elevation model
dsm   <- download_dem_kartverket(stn.id, stn.centre, name="dom", dx=100, resx=1)

# Compute the horizon
compute_horizon(stn.centre, dsm)
compute_horizon(stn.centre, dsm, level=stn.level, step=1, f.plot.polygon=TRUE)

}
\references{
\url{https://grass.osgeo.org/grass83/manuals/r.horizon.html}
}
