% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_horizon_max.R
\name{compute_horizon_max}
\alias{compute_horizon_max}
\title{Maximum horizon height around a station}
\usage{
compute_horizon_max(
  stn = NULL,
  step = 10,
  f_plot_polygon = FALSE,
  f_output_all = FALSE
)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_latlon_frost"}}

\item{step}{An interval in degrees at which the horizon will be computed,
default is every 10 deg.}

\item{f_plot_polygon}{A boolean setting boundary values to -20 deg to
facilitate plotting as polygon}
}
\value{
A dataframe with `horizon_height` in degrees and `azimuth` angle in
       degrees at which the horizon is computed in degrees
}
\description{
Compute maximum horizon height in degrees from a station location in UTM 33
and based on three DEMs: Digital Terrain Model, Digital Surface Model and
20km-Digital Terrain Model
}
\examples{
# Load the station metadata including location and level
stn <- get_metadata_frost(stationid = 18700, dx = 100, resx = 1)

# Compute the maximum horizon
compute_horizon_max(stn)
compute_horizon_max(stn, step = 1, f_plot_polygon = FALSE)

}
\references{
\url{https://grass.osgeo.org/grass83/manuals/r.horizon.html}
}
