% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_skyviewfactor.R
\name{compute_skyviewfactor}
\alias{compute_skyviewfactor}
\title{Sky View Factor (SVF)}
\usage{
compute_skyviewfactor(horizon = NULL)
}
\arguments{
\item{horizon}{An array of horizon height in degrees from \code{"compute_horizon"}}
}
\value{
Estimated sky view factor, 1 defines an open sky view and 0 a totally obstructed sky view
}
\description{
Compute Sky View Factor from horizon data
}
\examples{
compute_skyviewfactor(horizon)

}
\references{
\url{https://github.com/OSGeo/grass-addons/blob/grass8/src/raster/r.skyview/r.skyview.py}
}
