% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sun_position.R
\name{compute_sun_position}
\alias{compute_sun_position}
\title{Sun position in the sky}
\usage{
compute_sun_position(stn = NULL, f.hour = FALSE)
}
\arguments{
\item{stn}{A SpatVector with station attributes from \code{"get_latlon_frost"}}

\item{f.hour}{A boolean value to compute sun position for hours if TRUE and for days if FALSE}
}
\value{
dataframe with inclination of the sun position in the sky in degrees, azimuth in degrees and timestamp as factor
}
\description{
Compute sun position in the sky from a station location in UTM 33
}
\examples{
# Set parameters
stationid <- 18700
paramid <- 211

# Load the station metadata including location and level
stn <- get_latlon_frost(stationid,paramid)

# Compute sun position in the sky
compute_sun_position(stn, f.hour = TRUE)
compute_sun_position(stn, f.hour = FALSE)

}
\references{
\url{https://github.com/adokter/suntools/}
}
