% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_pagination.R
\name{extract_resp}
\alias{extract_resp}
\title{Loop through request responses from pagination}
\usage{
extract_resp(resp)
}
\arguments{
\item{resp}{A request response}
}
\value{
A matrix
}
\description{
Loop through request responses from pagination
}
\examples{
#No example
}
