% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latlon_frost.R
\name{get_latlon_frost}
\alias{get_latlon_frost}
\title{Get station and sensor metadata from Frost API}
\usage{
get_latlon_frost(stationid = 18700, paramid = NULL)
}
\arguments{
\item{stationid}{A station number as integer and defined by met.no, default is 18700 - Blindern}

\item{paramid}{A parameter number as integer and defined by met.no, default is NULL, thus getting all paramid available}
}
\value{
Station metadata
}
\description{
Fetch station metadata based on station number and parameter id from Frost v1.
The function gets station name, location and sensor details such as level, exposure and performance
}
\examples{
get_latlon_frost(stationid=18700)
get_latlon_frost(stationid=18700,paramid=211)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
