% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tile_wms.R
\name{get_tile_wms}
\alias{get_tile_wms}
\title{Get Web Mapping Service data tiles}
\usage{
get_tile_wms(box = NULL, layer = "ar5", px = 500)
}
\arguments{
\item{box}{A SpatExtent defining the area to plot}

\item{layer}{A name of a layer provided by WMS such as "ar5", "toporaster",
"CORINE_Land_Cover_2012", "Urban_Atlas_Lu_Lc_2012","ortofoto",
"fkb_arealdekke", "fkb_arealdekke","fkb_vann","vann_omrade","fkb_samferdsel",
"veg","fkb_bygning","bygning","fkb_naturinfo", "naturinfo"}

\item{px}{A pixel number that defines the resolution of the image/tile, default is 500 px}
}
\value{
A map tile
}
\description{
Fetch map tiles from publicly available WMS released by Nibio
}
\examples{
require(sf)

# Get station coordinates and name
stn    <- get_latlon_frost(18700)
centre <- sf::st_coordinates(stn)

# Construct box to extract WMS tile
dx <- 100
box <- c(c(centre[1],centre[2])-dx, c(centre[1],centre[2])+dx) \%>\% round()
class(box) <- "bbox"
box <- st_as_sfc(box)
st_crs(box) <- 25833 # UTM33

# Load tiles
tile <- get_tile_wms(box, layer = "ar5")
tile <- get_tile_wms(box, layer = "CORINE_Land_Cover_2012" )
tile <- get_tile_wms(box, layer = "Urban_Atlas_Lu_Lc_2012" )

}
\references{
\url{https://nibio.no/tjenester/wms-tjenester}
}
