% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata.R
\name{json_to_df}
\alias{json_to_df}
\title{Convert JSON to dataframe}
\usage{
json_to_df(json)
}
\arguments{
\item{json}{A JSON object}
}
\value{
A data.frame
}
\description{
Unlist JSON to matrix, then transpose and convert to dataframe
}
\examples{
#No example

}
