% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_make_bbox.R
\name{make_bbox}
\alias{make_bbox}
\title{Construct boundary box}
\usage{
make_bbox(stn = NULL, dx = NULL)
}
\arguments{
\item{stn}{A SpatVector or an array of two coordinates in UTM 33 (epsg:25833)}

\item{dx}{A distance in metre to compute the box boundary as a distance
from the centre point}
}
\value{
A bbox object
}
\description{
Construct boundary box from sf package
}
\examples{
# Get station metadata
stn <- get_metadata_frost(stationid = 18700)

# Define boundary box from SpatVector
box <- make_bbox(stn, dx = 1600)

# Get coordinates matrix and define boundary box
centre <- terra::crds(stn)
box <- make_bbox(centre, dx = 1600)

}
