% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dem_rayshader.R
\name{plot_dem_rayshader}
\alias{plot_dem_rayshader}
\title{Plot digital surface/elevation model in 3D with shadow rendering}
\usage{
plot_dem_rayshader(stn = NULL, dsm = NULL, path = NULL)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_latlon_frost"}}

\item{dsm}{A SpatRaster of a digital surface model around the station}

\item{path}{A string path that defines where to save the plot, if NULL
(default) the plot is printed on-screen and not saved}
}
\value{
A rendered image
}
\description{
Use rayshader to plot surface elevation model near a weather station from
four cardinal angles
}
\examples{
# Load the station metadata and location
stn <- get_metadata_frost(stationid = 18700, paramid = 211)

# Load a digital elevation model
dsm   <- download_dem_kartverket(stn, name = "dom", dx = 100, resx = 1)

# Plot 3D DEM with rayshader
plot_dem_rayshader(stn, dsm, path = 'plot/dem3D')

}
\references{
\url{https://www.rayshader.com/}
}
