% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_horizon_sun.R
\name{plot_station_horizon_sun}
\alias{plot_station_horizon_sun}
\title{Plot horizon with sun positions for a given station location}
\usage{
plot_station_horizon_sun(
  stn = NULL,
  dem = NULL,
  dsm = NULL,
  demkm = NULL,
  path = NULL
)
}
\arguments{
\item{stn}{A SpatVector with station attributes from \code{"get_latlon_frost"}}

\item{dem}{A SpatRaster of a digital elevation/terrain model around the station, expected radius is 100 m}

\item{dsm}{A SpatRaster of a digital surface model around the station, expected radius is 100 m}

\item{demkm}{A SpatRaster of a digital elevation/terrain model around the station, expected radius is 20 km}

\item{path}{A directory path defining where will be saved the plots, if path is NULL the plots are printed to the console}
}
\value{
Sun diagram with station metadata
}
\description{
Fetch station metadata based on station number and parameter id from Frost v1.
The function gets station name, location and sensor details such as level, exposure and performance
}
\examples{
require(sf)

# Load the station metadata and location
stn <- get_latlon_frost(stationid=18700, paramid=211)
stn.id      <- stn$id.stationid
stn.centre  <- stn  \%>\% st_coordinates()

# Load DEM data
dem   <- download_dem_kartverket(stn.id, stn.centre, name="dtm",dx=100,resx=1)
dsm   <- download_dem_kartverket(stn.id, stn.centre, name="dom",dx=100,resx=1)
demkm <- download_dem_kartverket(stn.id, stn.centre, name="dtm",dx=20e3,resx=20)

# Plot sun diagram and save
path <- 'plot/horizon'
plot_station_horizon_sun(stn, dem, dsm, demkm, path=path)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
