% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_horizon_sun.R
\name{plot_station_horizon_sun}
\alias{plot_station_horizon_sun}
\title{Plot horizon with sun positions for a given station location}
\usage{
plot_station_horizon_sun(
  stn = NULL,
  dem = NULL,
  dsm = NULL,
  demkm = NULL,
  path = NULL
)
}
\arguments{
\item{stn}{A SpatVector with station attributes from \code{"get_latlon_frost"}}

\item{dem}{A SpatRaster of a digital elevation/terrain model around the station, expected radius is 100 m}

\item{dsm}{A SpatRaster of a digital surface model around the station, expected radius is 100 m}

\item{demkm}{A SpatRaster of a digital elevation/terrain model around the station, expected radius is 20 km}

\item{path}{A directory path defining where will be saved the plots, if path is NULL the plots are printed to the console}
}
\value{
Sun diagram with station metadata
}
\description{
Fetch station metadata based on station number and parameter id from Frost v1.
The function gets station name, location and sensor details such as level, exposure and performance
}
\examples{
# Load data
stn <- get_latlon_frost(stationid=18700)
dem   <- download_dem_kartverket(stationid,centre,name="dtm",dx=100,resx=1)
dsm   <- download_dem_kartverket(stationid,centre,name="dom",dx=100,resx=1)
demkm <- download_dem_kartverket(stationid,centre,name="dtm",dx=20e3,resx=20)
path <- sprintf("station_location_files/output/\%i",stn$id.stationid)
# path <- 'plot/horizon'
# Plot sun diagram
plot_station_horizon_sun(stn, dem, dsm, demkm, path=path)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
