% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_horizon_sun.R
\name{plot_station_horizon_sun}
\alias{plot_station_horizon_sun}
\title{Plot horizon with sun positions for a given station location}
\usage{
plot_station_horizon_sun(
  stn = NULL,
  dem = NULL,
  dsm = NULL,
  demkm = NULL,
  path = NULL
)
}
\arguments{
\item{stn}{A SpatVector with station attributes from
\code{"get_latlon_frost"}}

\item{dem}{A SpatRaster of a digital terrain model around the station}

\item{dsm}{A SpatRaster of a digital surface model around the station}

\item{demkm}{A SpatRaster of a wider digital terrain model around the station}

\item{path}{A directory path defining where will be saved the plots,
if path is NULL the plots are printed to the console}
}
\value{
Sun diagram with station metadata
}
\description{
Plot sun position and horizons from local terrain model, local surface model
and regional digital terrain model with infos from the weather station.
}
\examples{
# Load the station metadata and location
stn <- get_latlon_frost(stationid = 18700, paramid = 211)

# Load DEM data
dem   <- download_dem_kartverket(stn, name = "dtm", dx = 100, resx = 1)
dsm   <- download_dem_kartverket(stn, name = "dom", dx = 100, resx = 1)
demkm <- download_dem_kartverket(stn, name = "dtm", dx = 20e3, resx = 20)

# Plot sun diagram and save
path <- 'plot/horizon'
plot_station_horizon_sun(stn, dem, dsm, demkm, path = path)

}
\references{
\url{https://frost-beta.met.no/docs/codeexamples}
}
