% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_siting_context.R
\name{plot_station_siting_context}
\alias{plot_station_siting_context}
\title{Weather station context}
\usage{
plot_station_siting_context(
  stationid = 18700,
  paramid = 211,
  f_verbose = FALSE,
  f_pdf = FALSE
)
}
\arguments{
\item{stationid}{A station number as integer and defined by met.no}

\item{paramid}{A parameter number as integer and defined by met.no}

\item{f_verbose}{A boolean string to print debug messages, default is TRUE}

\item{f_pdf}{A boolean string to combine all plots into a pdf file}
}
\value{
None
}
\description{
Plots weather station's sun diagram and background maps to assess
the exposure of a station and to compute its WMO/met.no siting classification
}
\examples{
# Plot sun diagram and map infos for a weather station
plot_station_siting_context(stationid = 18700)
#plot_station_siting_context(stationid = 18700,
 #                           paramid = 211,
  #                          f_verbose = TRUE)

}
\references{
\url{https://community.wmo.int/en/activity-areas/
imop/siting-classification#MembersTools}
}
